cleanup rspec context, fixes
