+++ drivers/hud_spm_report/spec/models/fy2024/shared_context.rb
@@ -1,9 +1,11 @@
 # frozen_string_literal: true
 
 require 'rails_helper'
+require_relative '../../../../../spec/shared_contexts/hud_enrollment_builders'
 
 # Shared context for SPM testing
 RSpec.shared_context 'SPM test setup', shared_context: :metadata do
+  include_context 'HUD enrollment builders'
   let(:user) { create(:user) }
   let(:default_filter) do
     Filters::HudFilterBase.new(
@@ -20,70 +22,6 @@ RSpec.shared_context 'SPM test setup', shared_context: :metadata do
 
   # Setup CoC organization
   let!(:organization) { create(:hud_organization, data_source: data_source) }
-
-  def create_project(project_type:, coc_code: 'MA-500')
-    project = create(
-      :hud_project,
-      project_type: project_type,
-      organization: organization,
-      data_source: data_source,
-      ContinuumProject: 1,
-    )
-
-    create(
-      :hud_project_coc,
-      project_id: project.project_id,
-      data_source: data_source,
-      coc_code: coc_code,
-    )
-
-    project
-  end
-
-  def create_client_with_warehouse_link(dob: '1995-04-05'.to_date)
-    client = create(:hud_client, data_source: data_source, dob: dob)
-    destination_client = create(:hud_client, data_source: destination_data_source)
-    create(:warehouse_client, destination_id: destination_client.id, source_id: client.id)
-    client
-  end
-
-  def create_enrollment(client:, project:, entry_date:, exit_date: nil, relationship_to_ho_h: 1, date_to_street_essh: nil, household_id: Hmis::Hud::Base.generate_uuid, living_situation: nil, destination: nil, move_in_date: nil)
-    enrollment = create(
-      :hud_enrollment,
-      client: client,
-      project: project,
-      data_source: data_source,
-      entry_date: entry_date,
-      date_to_street_essh: date_to_street_essh,
-      relationship_to_ho_h: relationship_to_ho_h,
-      household_id: household_id,
-      living_situation: living_situation,
-      move_in_date: move_in_date,
-    )
-
-    if exit_date.present?
-      create(
-        :hud_exit,
-        enrollment: enrollment,
-        exit_date: exit_date,
-        data_source: data_source,
-        personal_id: client.personal_id,
-        destination: destination,
-      )
-    end
-
-    enrollment
-  end
-
-  def create_bed_night_service(enrollment:, date:)
-    create(
-      :hud_service,
-      enrollment: enrollment,
-      date_provided: date,
-      data_source: data_source,
-      record_type: 200, # bed night
-    )
-  end
 
   def setup_report(project_ids, questions = ['Measure 1'])
     filter = default_filter.dup
